Refactored MainScreen into own file + Implement Button functionality
MainScreen moved into own file for readability and convention.
setEffect() now creates a file in pmeas-backend/config that is filled with a test effect to be
used for implementing backend system.
